Update Profile.jsx
Updates to profile page. This is not functional yet, do not merge to main. I have the profile page pulling the users first name, last name, and email address from the database information from when the user first registered. The other information is set to default values. When the user updates the information it should save to the database. The user should also receive an email with what information has been updated in their profile. When I go to the profile page now, the data fields are not loading, so I don't know if that means it is having trouble pulling the data from the database or if I just have it coded wrong. I would appreciate some help, because I still mostly have no idea what I am doing.